# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Invite Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface AuthUser {
  id: string;
  username: string;
  discriminator: string;
  /** @nullable */
  avatar?: string | null;
  avatarUrl: string;
}

export interface Guild {
  id: string;
  name: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  iconUrl?: string | null;
  premiumTier: number;
  /** @nullable */
  memberCount?: number | null;
  isAdmin: boolean;
}

export interface GuildDetail {
  id: string;
  name: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  iconUrl?: string | null;
  premiumTier: number;
  /** @nullable */
  memberCount?: number | null;
  isAdmin: boolean;
  totalInvites: number;
  totalMembers: number;
  activeTrackers: number;
}

export interface GuildSettings {
  guildId: string;
  /** @nullable */
  joinMessage?: string | null;
  /** @nullable */
  leaveMessage?: string | null;
  /** @nullable */
  joinChannelId?: string | null;
  /** @nullable */
  leaveChannelId?: string | null;
  fakeThreshold: number;
  trackFakeInvites: boolean;
  trackLeaves: boolean;
  /** @nullable */
  embedColor?: string | null;
  /** @nullable */
  embedImage?: string | null;
}

export interface UpdateGuildSettingsBody {
  /** @nullable */
  joinMessage?: string | null;
  /** @nullable */
  leaveMessage?: string | null;
  /** @nullable */
  joinChannelId?: string | null;
  /** @nullable */
  leaveChannelId?: string | null;
  /** @nullable */
  fakeThreshold?: number | null;
  /** @nullable */
  trackFakeInvites?: boolean | null;
  /** @nullable */
  trackLeaves?: boolean | null;
  /** @nullable */
  embedColor?: string | null;
  /** @nullable */
  embedImage?: string | null;
}

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  username: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  regular: number;
  bonus: number;
  fake: number;
  left: number;
  total: number;
}

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
  total: number;
  page: number;
  limit: number;
}

export interface TrendPoint {
  date: string;
  joins: number;
  leaves: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  join: "join",
  leave: "leave",
} as const;

export interface ActivityItem {
  type: ActivityItemType;
  userId: string;
  username: string;
  /** @nullable */
  inviterId?: string | null;
  /** @nullable */
  inviterName?: string | null;
  timestamp: string;
  isFake: boolean;
}

export interface GuildStats {
  totalJoins: number;
  totalLeaves: number;
  totalFake: number;
  netGrowth: number;
  /** @nullable */
  topInviter?: string | null;
  /** @nullable */
  topInviterCount?: number | null;
  joinsTrend: TrendPoint[];
  recentActivity: ActivityItem[];
}

export interface MemberJoin {
  id: number;
  userId: string;
  username: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  inviterId?: string | null;
  /** @nullable */
  inviterName?: string | null;
  /** @nullable */
  inviteCode?: string | null;
  joinedAt: string;
  /** @nullable */
  leftAt?: string | null;
  isFake: boolean;
  /** @nullable */
  accountAgeDays?: number | null;
}

export interface MembersResponse {
  members: MemberJoin[];
  total: number;
  page: number;
  limit: number;
}

export interface UserInviteStats {
  userId: string;
  username: string;
  /** @nullable */
  avatar?: string | null;
  regular: number;
  bonus: number;
  fake: number;
  left: number;
  total: number;
  /** @nullable */
  rank?: number | null;
  invitedMembers?: MemberJoin[];
}

export interface AdjustBonusBody {
  userId: string;
  amount: number;
  /** @nullable */
  reason?: string | null;
}

export type ResetInvitesBodyType =
  (typeof ResetInvitesBodyType)[keyof typeof ResetInvitesBodyType];

export const ResetInvitesBodyType = {
  all: "all",
  regular: "regular",
  bonus: "bonus",
  fake: "fake",
  left: "left",
} as const;

export interface ResetInvitesBody {
  userId: string;
  type: ResetInvitesBodyType;
}

export interface PremiumPlan {
  tier: number;
  name: string;
  price: number;
  currency: string;
  features: string[];
  highlighted: boolean;
}

export interface PremiumStatus {
  tier: number;
  tierName: string;
  /** @nullable */
  expiresAt?: string | null;
  isActive: boolean;
}

export type GetLeaderboardParams = {
  page?: number;
  limit?: number;
};

export type GetGuildMembersParams = {
  page?: number;
  limit?: number;
};