# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Invite Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /guilds:
    get:
      operationId: getGuilds
      tags: [guilds]
      summary: List guilds the bot is in and user has access to
      responses:
        "200":
          description: List of guilds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guild"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /guilds/{guildId}:
    get:
      operationId: getGuild
      tags: [guilds]
      summary: Get a specific guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Guild details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildDetail"
        "404":
          description: Guild not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /guilds/{guildId}/settings:
    get:
      operationId: getGuildSettings
      tags: [guilds]
      summary: Get guild settings
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Guild settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildSettings"
    put:
      operationId: updateGuildSettings
      tags: [guilds]
      summary: Update guild settings
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGuildSettingsBody"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildSettings"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    AuthUser:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        discriminator:
          type: string
        avatar:
          type: ["string", "null"]
        avatarUrl:
          type: string
      required:
        - id
        - username
        - discriminator
        - avatarUrl
    Guild:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: ["string", "null"]
        iconUrl:
          type: ["string", "null"]
        premiumTier:
          type: integer
        memberCount:
          type: ["integer", "null"]
        isAdmin:
          type: boolean
      required:
        - id
        - name
        - premiumTier
        - isAdmin
    GuildDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: ["string", "null"]
        iconUrl:
          type: ["string", "null"]
        premiumTier:
          type: integer
        memberCount:
          type: ["integer", "null"]
        isAdmin:
          type: boolean
        totalInvites:
          type: integer
        totalMembers:
          type: integer
        activeTrackers:
          type: integer
      required:
        - id
        - name
        - premiumTier
        - isAdmin
        - totalInvites
        - totalMembers
        - activeTrackers
    GuildSettings:
      type: object
      properties:
        guildId:
          type: string
        joinMessage:
          type: ["string", "null"]
        leaveMessage:
          type: ["string", "null"]
        joinChannelId:
          type: ["string", "null"]
        leaveChannelId:
          type: ["string", "null"]
        fakeThreshold:
          type: integer
        trackFakeInvites:
          type: boolean
        trackLeaves:
          type: boolean
        embedColor:
          type: ["string", "null"]
        embedImage:
          type: ["string", "null"]
      required:
        - guildId
        - fakeThreshold
        - trackFakeInvites
        - trackLeaves
    UpdateGuildSettingsBody:
      type: object
      properties:
        joinMessage:
          type: ["string", "null"]
        leaveMessage:
          type: ["string", "null"]
        joinChannelId:
          type: ["string", "null"]
        leaveChannelId:
          type: ["string", "null"]
        fakeThreshold:
          type: ["integer", "null"]
        trackFakeInvites:
          type: ["boolean", "null"]
        trackLeaves:
          type: ["boolean", "null"]
        embedColor:
          type: ["string", "null"]
        embedImage:
          type: ["string", "null"]
...